Rename `rtdb/variables.ts` to `rtdb/constants.ts`
